Database placed in shared module
