+++ src/django/Dockerfile
@@ -4,7 +4,8 @@ ARG BE_PORT
 
 RUN apt update -y && \
 	apt upgrade -y && \
-	sudo apt-get install libpq-dev \
+	apt install libpq-dev -y && \
+	apt install gdal-bin -y && \
 	pip3 install pipenv && \
 	apt -y autoremove && \
     apt -y clean
