+++ src/django/app/shared/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from .models import Student
 
 # Register your models here.
+admin.site.register(Student)
