+++ src/django/app/user/forms.py
@@ -1,12 +1,12 @@
-from .models import CustomUser
+from .models import Student
 from django.contrib.auth.forms import UserCreationForm, UserChangeForm
 
 class CustomUserCreationForm(UserCreationForm):
     class Meta:
-        model = CustomUser
+        model = Student
         fields = ['first_name', 'last_name', 'email']
 
 class CustomUserChangeForm(UserChangeForm):
     class Meta:
-        model = CustomUser
+        model = Student
         fields = ['first_name', 'last_name', 'email']
