+++ src/django/app/user/models.py
@@ -1,35 +1,3 @@
-from django.db import models
-from django.contrib.auth.models import (
-    AbstractBaseUser,
-    BaseUserManager,
-    PermissionsMixin,
-)
+from shared.models import *
+    
 
-
-class CustomUserManager(BaseUserManager):
-
-    def create_user(self, email, password, **extra_fields):
-        if not email:
-            raise ValueError('User must have an email address.')
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, email, password):
-        user = self.create_user(email, password)
-        user.is_staff = True
-        user.is_superuser = True
-        user.save()
-        return user
-class CustomUser(AbstractBaseUser, PermissionsMixin):
-
-    email = models.EmailField(max_length=255, unique=True)
-    first_name = models.CharField(max_length=255)
-    last_name = models.CharField(max_length=255)
-    is_staff = models.BooleanField(default=False)
-    is_active = models.BooleanField(default=True)
-
-    objects = CustomUserManager()
-
-    USERNAME_FIELD = 'email'
